# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ChattyCathy API
  description: |
    ChattyCathy API Server with JWT authentication and RBAC.

    ## Authentication

    This API uses JWT (RS256) authentication with access and refresh tokens.

    - **Access tokens**: Short-lived (15 min), used in Authorization header
    - **Refresh tokens**: Long-lived (7 days), used to obtain new access tokens

    ### Authentication Methods

    - **Google OAuth**: Sign in with Google (recommended for users)
    - **Username/Password**: Traditional login (for demo/testing)

    ### For Web Clients
    Tokens are automatically set as HTTP-only cookies after login.

    ### For Mobile/API Clients
    Use the `Authorization: Bearer <token>` header with access tokens.

    ## Role-Based Access Control (RBAC)

    The API implements permission-based access control:

    ### Default Roles
    - **admin**: Full access to all resources
    - **user**: Basic read access (ping, news) - *automatically assigned to new users*
    - **editor**: Read and write access to content

    ### Default Permissions
    - `ping:read` - Access ping endpoint
    - `news:read`, `news:create`, `news:update`, `news:delete`
    - `users:read`, `users:create`, `users:update`, `users:delete`
    - `roles:read`, `roles:create`, `roles:update`, `roles:delete`

    ### Automatic Role Assignment
    New users registered via Google OAuth are automatically assigned the **user** role,
    which grants `ping:read` and `news:read` permissions.

...
servers:
...
paths:
...
  /admin/permissions:
    get:
      summary: List all permissions
      description: Returns all available permissions in the system. Requires admin role.
      operationId: listPermissions
      tags:
        - admin
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Permission"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - requires admin role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/roles:
    get:
      summary: List all roles
      description: Returns all roles with their associated permissions. Requires admin role.
      operationId: listRoles
      tags:
        - admin
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - requires admin role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Create a new role
      description: Creates a new role. Requires admin role.
      operationId: createRole
      tags:
        - admin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleRequest"
      responses:
        "201":
          description: Role created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - requires admin role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Role already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/roles/{id}:
    get:
      summary: Get a role
      description: Returns a single role with its permissions. Requires admin role.
      operationId: getRole
      tags:
        - admin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Role ID
      responses:
        "200":
          description: Role details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - requires admin role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Role not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      summary: Update a role
      description: Updates a role's name and description. System roles cannot be renamed. Requires admin role.
      operationId: updateRole
      tags:
        - admin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Role ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleRequest"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - requires admin role or trying to rename system role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Role not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Role name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete a role
      description: Deletes a role. System roles cannot be deleted. Requires admin role.
      operationId: deleteRole
      tags:
        - admin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Role ID
      responses:
        "200":
          description: Role deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: role deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - requires admin role or trying to delete system role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Role not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/roles/{id}/permissions:
    put:
      summary: Set role permissions
      description: |
        Replaces all permissions for a role with the provided list.
        Requires admin role.
      operationId: setRolePermissions
      tags:
        - admin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Role ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetRolePermissionsRequest"
      responses:
        "200":
          description: Permissions updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - requires admin role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Role not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Permission:
      type: object
      properties:
        id:
          type: integer
          description: Permission ID
          example: 1
        name:
          type: string
          description: Permission name in format resource:action
          example: "news:read"
        description:
          type: string
          description: Human-readable description
          example: "Can read news articles"
        resource:
          type: string
          description: Resource this permission applies to
          example: "news"
        action:
          type: string
          description: Action this permission grants
          example: "read"
      required:
        - id
        - name
        - description
        - resource
        - action
    Role:
      type: object
      properties:
        id:
          type: integer
          description: Role ID
          example: 1
        name:
          type: string
          description: Role name
          example: "editor"
        description:
          type: string
          description: Human-readable description
          example: "Editor with news management access"
        is_system:
          type: boolean
          description: Whether this is a system role (cannot be deleted)
          example: false
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
          description: Permissions assigned to this role
      required:
        - id
        - name
        - description
        - is_system
        - permissions
    CreateRoleRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          description: Role name
          example: "moderator"
        description:
          type: string
          maxLength: 255
          description: Role description
          example: "Moderator with content moderation access"
      required:
        - name
    UpdateRoleRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          description: Role name
          example: "moderator"
        description:
          type: string
          maxLength: 255
          description: Role description
          example: "Updated moderator description"
      required:
        - name
    SetRolePermissionsRequest:
      type: object
      properties:
        permission_ids:
          type: array
          items:
            type: integer
          description: List of permission IDs to assign to the role
          example: [1, 2, 3]
      required:
        - permission_ids